+++ lib/game/big_brother_game.dart
@@ -1,7 +1,5 @@
 import 'package:big_brother/entities/background/background_entity.dart';
 import 'package:big_brother/entities/level/level_entity.dart';
-import 'package:big_brother/entities/ui/input_button.dart';
-import 'package:big_brother/entities/ui/input_button_icon.dart';
 import 'package:flame/game.dart';
 import 'package:flame/input.dart';
 import 'package:flutter/material.dart';
@@ -9,6 +7,7 @@ import 'package:flutter/material.dart';
 class BigBrotherGame extends FlameGame
     with HasCollisionDetection, HasKeyboardHandlerComponents {
   final Paint _backgroundColor = Paint()..color = const Color(0xff211f30);
+  final level = LevelEntity();
 
   @override
   Future<void> onLoad() async {
@@ -16,8 +15,7 @@ class BigBrotherGame extends FlameGame
     camera.viewfinder.zoom = 1.5;
 
     world.add(BackgroundEntity());
-    world.add(LevelEntity());
-    world.add(InputButton(icon: InputButtonIcon.keyE, duration: 3));
+    world.add(level);
   }
 
   @override